rename project keygenerator to secretgenerator across module path binary container brand and docs
