+++ Dockerfile
@@ -3,8 +3,8 @@ FROM gcr.io/distroless/static-debian12:nonroot
 ARG TARGETOS
 ARG TARGETARCH
 
-COPY keygenerator /usr/local/bin/keygenerator
+COPY secretgenerator /usr/local/bin/secretgenerator
 
 USER nonroot:nonroot
 
-ENTRYPOINT ["/usr/local/bin/keygenerator"]
+ENTRYPOINT ["/usr/local/bin/secretgenerator"]
