+++ cmd/secretgenerator/entropy_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/rafaelperoco/keygenerator/internal/audit"
+	"github.com/rafaelperoco/secretgenerator/internal/audit"
 )
 
 func baseEntropyOptions(stdout, stderr io.Writer) entropyOptions {
