+++ cmd/secretgenerator/exit.go
