+++ cmd/secretgenerator/version.go
