+++ docs/AUDIT.md
@@ -1,6 +1,6 @@
-# Auditing keygenerator releases
+# Auditing secretgenerator releases
 
-This document describes how to verify, end-to-end, that a `keygenerator` binary
+This document describes how to verify, end-to-end, that a `secretgenerator` binary
 you have downloaded was built from the source in this repository, by GitHub
 Actions, with no tampering in transit.
 
@@ -13,7 +13,7 @@ You ─trust─▶ Sigstore Fulcio (root CA)
                                   │
                                   └─signs─▶ checksums.txt
                                           │
-                                          └─pins SHA-256 of─▶ keygenerator binary
+                                          └─pins SHA-256 of─▶ secretgenerator binary
 ```
 
 You do **not** need to trust this README, the project maintainer, or any web
@@ -21,7 +21,7 @@ host. You only need to trust:
 
 1. The Sigstore project's root keys (distributed with `cosign`).
 2. GitHub Actions' OIDC token issuer (a known, audited service).
-3. That the GitHub repository at `github.com/rafaelperoco/keygenerator` is the
+3. That the GitHub repository at `github.com/rafaelperoco/secretgenerator` is the
    one you intend to install from.
 
 Everything else is verifiable cryptographically.
@@ -44,26 +44,26 @@ brew install syft
 
 ### 1. Download the artifacts
 
-From the [release page](https://github.com/rafaelperoco/keygenerator/releases),
+From the [release page](https://github.com/rafaelperoco/secretgenerator/releases),
 download:
 
-- `keygenerator_<version>_<os>_<arch>.tar.gz` — the binary archive
+- `secretgenerator_<version>_<os>_<arch>.tar.gz` — the binary archive
 - `checksums.txt` — SHA-256 hashes of every release artifact
 - `checksums.txt.sig` — cosign signature over `checksums.txt`
 - `checksums.txt.pem` — cosign certificate (Fulcio-issued)
 - `multiple.intoto.jsonl` — SLSA Level 3 provenance attestation
-- `keygenerator_<version>_<os>_<arch>.tar.gz.sbom.json` — CycloneDX SBOM
+- `secretgenerator_<version>_<os>_<arch>.tar.gz.sbom.json` — CycloneDX SBOM
 
 ### 2. Verify the cosign signature on `checksums.txt`
 
 This step proves: "the GitHub Actions workflow at
-`github.com/rafaelperoco/keygenerator` produced this `checksums.txt`."
+`github.com/rafaelperoco/secretgenerator` produced this `checksums.txt`."
 
 ```sh
 cosign verify-blob \
   --certificate checksums.txt.pem \
   --signature checksums.txt.sig \
-  --certificate-identity-regexp "https://github.com/rafaelperoco/keygenerator/.github/workflows/release.yml@refs/tags/v.*" \
+  --certificate-identity-regexp "https://github.com/rafaelperoco/secretgenerator/.github/workflows/release.yml@refs/tags/v.*" \
   --certificate-oidc-issuer "https://token.actions.githubusercontent.com" \
   checksums.txt
 ```
@@ -78,7 +78,7 @@ This step proves: "the binary I have matches what the workflow signed."
 sha256sum -c checksums.txt --ignore-missing
 ```
 
-Expected output: `keygenerator_<version>_<os>_<arch>.tar.gz: OK`.
+Expected output: `secretgenerator_<version>_<os>_<arch>.tar.gz: OK`.
 
 ### 4. Verify SLSA Level 3 provenance
 
@@ -88,9 +88,9 @@ in a hardened CI environment, by a non-falsifiable workflow."
 ```sh
 slsa-verifier verify-artifact \
   --provenance-path multiple.intoto.jsonl \
-  --source-uri github.com/rafaelperoco/keygenerator \
+  --source-uri github.com/rafaelperoco/secretgenerator \
   --source-tag v<version> \
-  keygenerator_<version>_<os>_<arch>.tar.gz
+  secretgenerator_<version>_<os>_<arch>.tar.gz
 ```
 
 Expected output: `Verified signature against tlog entry ...` and
@@ -102,18 +102,18 @@ After extracting the archive, the binary itself reports its build identity in
 every JSON output. Check that it matches the release tag:
 
 ```sh
-./keygenerator --json | jq '{version, commit, build_date}'
+./secretgenerator --json | jq '{version, commit, build_date}'
 ```
 
 The `version` should match the release tag, and `commit` should match the git
-SHA shown on the [release page](https://github.com/rafaelperoco/keygenerator/releases).
+SHA shown on the [release page](https://github.com/rafaelperoco/secretgenerator/releases).
 
 ### 6. (Optional) Inspect the SBOM
 
 The SBOM lists every dependency that went into the build:
 
 ```sh
-syft scan keygenerator_<version>_<os>_<arch>.tar.gz.sbom.json
+syft scan secretgenerator_<version>_<os>_<arch>.tar.gz.sbom.json
 ```
 
 You can also scan for known vulnerabilities. The release pipeline already
@@ -121,20 +121,20 @@ runs Trivy as a gate before publishing, but a fresh scan from your machine
 catches CVEs disclosed after the release was cut:
 
 ```sh
-trivy sbom keygenerator_<version>_<os>_<arch>.tar.gz.sbom.json
+trivy sbom secretgenerator_<version>_<os>_<arch>.tar.gz.sbom.json
 # Alternative tool with same SBOM input
-grype sbom:keygenerator_<version>_<os>_<arch>.tar.gz.sbom.json
+grype sbom:secretgenerator_<version>_<os>_<arch>.tar.gz.sbom.json
 ```
 
 The project's own daily Trivy scan results are visible in the GitHub
 Security tab at
-https://github.com/rafaelperoco/keygenerator/security/code-scanning
+https://github.com/rafaelperoco/secretgenerator/security/code-scanning
 
 ## Verifying a container image
 
 ```sh
-cosign verify ghcr.io/rafaelperoco/keygenerator:v<version> \
-  --certificate-identity-regexp "https://github.com/rafaelperoco/keygenerator/.github/workflows/release.yml@refs/tags/v.*" \
+cosign verify ghcr.io/rafaelperoco/secretgenerator:v<version> \
+  --certificate-identity-regexp "https://github.com/rafaelperoco/secretgenerator/.github/workflows/release.yml@refs/tags/v.*" \
   --certificate-oidc-issuer "https://token.actions.githubusercontent.com"
 ```
 
@@ -151,5 +151,5 @@ cosign verify ghcr.io/rafaelperoco/keygenerator:v<version> \
 ## Reporting verification failures
 
 If any step above fails, **do not run the binary**. Open a security advisory
-at https://github.com/rafaelperoco/keygenerator/security/advisories/new with
+at https://github.com/rafaelperoco/secretgenerator/security/advisories/new with
 the artifact filenames, the exact command you ran, and the failure output.
