+++ pkg/secretgen/secretgen_test.go
@@ -1,4 +1,4 @@
-package keygen
+package secretgen
 
 import (
 	"errors"
